+++ ad_header.php
@@ -13,16 +13,16 @@ if ($main_var != 'parol') exit;     // защита от запуска этог
     <link type="text/css" href="css/jquery-ui.css" rel="stylesheet" />
     <link type="text/css" href="css/style.css" rel="stylesheet" />
     <!-- include jquery libs -->
-    <script src="js/jquery.min.js" type="text/javascript"></script>
+
+    <script src="//code.jquery.com/jquery-1.11.1.min.js" type="text/javascript"></script>
     <script type="text/javascript" src="./js/jquery.form.js"></script>
-    <script src="js/jquery-ui.custom.min.js" type="text/javascript"></script>
+    <script src="//code.jquery.com/ui/1.11.2/jquery-ui.min.js" type="text/javascript"></script>
     <script type="text/javascript" src="js/jquery.cookie.js"></script>
     <!-- include column resize lib -->
-    <script src="js/colResizable-1.3.med.js" type="text/javascript"></script>
+    <script src="js/colResizable-1.5.min.js" type="text/javascript"></script>
     <!-- fancytree -->
     <link type="text/css" href="fancytree/skin-win8/ui.fancytree.css" rel="stylesheet" />
     <script src="fancytree/jquery.fancytree-all.js" type="text/javascript"></script>
-    <script type="text/javascript" src="js/jquery.fancytree.persist.js"></script>
     <!-- /fancytree -->
     <script type="text/javascript" src="js/jscript.js"></script>
     <title>AD FrontEnd on <?php echo "$site_name" ?></title>